+++ install.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
-set -ex
 
 # This script will create a systemd unit for running CLiC uWSGI, and
 # an nginx config.
@@ -7,7 +6,4 @@ set -ex
 # It is tested on Debian, but should hopefully work on anything systemd-based.
 #
 # See each part for detail on what happens
-
-. schema/deploy/deploy.sh
-. server/deploy/deploy.sh
-. client/deploy/deploy.sh
+exec make install